Main Game Loops Fixed
The main game loop was broken and would not recognize when the Hand was
over (player or dealer lost/won). When someone lost or won the game was
supposed to prompt the user to play again.

The error was in Game.java line: 61 where the user would call
dealer.firstDeal() when prompted to play instead of calling play() from
the Game class.

The boolean variable to identify when the Hand was over was moved from
the Player Class to the Dealer Class to reduce instances of error and
method calls.

Still have a bug where the loop in dealerHit() doesn't execute when the
dealer's visible card is an Ace/the player purchases insurance. Cause
not identified yet, but reproducable.